+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Internal
 
+- Fix CFS metric expression in the FeatureSelection class.
 - Fix the vcpkg configuration in building the library.
 - Fix the sample app to use the vcpkg configuration.
 - Add predict_proba method to all Ld classifiers.